Move segyio/_header.py -> segyio/header.py
In preparation for a better, more well-documentated and cleaner
implementation of the header mode, move it to not be _-prefixed.
