+++ python/segyio/header.py
